parseFloat()

+++ second.js
@@ -2,7 +2,7 @@
 //To ensure we have integer input, we use parseInt()
 const prompt = require("prompt-sync")();
 
-const age = parseInt(prompt("Enter you age: "));
+const age = parseFloat(prompt("Enter you age: "));
 console.log('You are', age); //but prompt saves input as strings
 
 console.log(typeof(age));
